v0.9.5

+++ client.lua
@@ -4,6 +4,7 @@ local messages = {}
 local currentContact = 1
 local currentMessage = 1
 local currentPagerObj = 0
+local isPagerActive = false
 
 --------------------------------------------------------------------------
 -- Gets pager data from server
@@ -109,6 +110,7 @@ end
 --------------------------------------------------------------------------
 AddEventHandler('QBCore:Client:OnPlayerLoaded', function() 
     GetPagerData()
+    -- request animations
     lib.requestAnimDict(Config.Animations.usePager.dict)
     lib.requestAnimDict(Config.Animations.getPagerOutOfPocket.dict)
     lib.requestAnimDict(Config.Animations.putPagerInPocket.dict)
@@ -148,6 +150,8 @@ end)
 --  when the pager is used)
 --------------------------------------------------------------------------
 RegisterNetEvent("96rp-pager:pager:show", function()
+    isPagerActive = true
+    -- show ui
     SetNuiFocus(true, true)
     SendNUIMessage({
         text = string.format("Welcome :)<br> Your number: %s", number),
@@ -155,22 +159,23 @@ RegisterNetEvent("96rp-pager:pager:show", function()
     })
     currentMessage = #messages + 1
 
+    -- trigger getPagerOutOfPocket animation
     local playerPed = PlayerPedId()
-    if not IsPedInAnyVehicle(playerPed, false) then
-        TriggerServerEvent('96rp-pager:server:PlayAnimation', true)
-    end
+    TriggerServerEvent('96rp-pager:server:PlayAnimation', 'getPagerOutOfPocket')
     while not IsEntityPlayingAnim(playerPed, Config.Animations.getPagerOutOfPocket.dict, Config.Animations.getPagerOutOfPocket.name, 3) do
-        Wait(100)
+        Wait(0)
     end
     Wait(1000)
-    local playerCoords = GetEntityCoords(playerPed)
     
+    -- delete last pager (if player spamms animation)
     if DoesEntityExist(currentPagerObj) then
         DeleteEntity(currentPagerObj)
     end
+
+    -- attach entity to players hand
+    local playerCoords = GetEntityCoords(playerPed)
     currentPagerObj = CreateObject(Config.PagerObj, playerCoords.x, playerCoords.y, playerCoords.z + 3, true, false, false)
     local boneIndex = GetPedBoneIndex(playerPed, 28422)
-
     local x = 0.0
     local y = 0.0
     local z = 0.0
@@ -184,23 +189,31 @@ RegisterNetEvent("96rp-pager:pager:show", function()
     local rotationorder = 2
     local syncRotation = true
     AttachEntityToEntity(currentPagerObj, playerPed, boneIndex, x, y, z, rotX, rotY, rotZ, p9, useSoftPinning, collision, isPed, rotationorder, syncRotation)
+    
+    -- wait for getPagerOutOfPocket animation to stop
+    local timeLeft = Config.Animations.getPagerOutOfPocket.time / 20
+    while timeLeft > 0 and isPagerActive do
+        timeLeft = timeLeft - 1
+        Wait(1)
+    end
+
+    -- timeLeft wont be <=0 if player spam-open-closes the pager and the usePager animation wont play
+    if timeLeft <= 0 then
+        TriggerServerEvent('96rp-pager:server:PlayAnimation', 'usePager')
+    end
 end)
 
 --------------------------------------------------------------------------
 -- Closes pager
 --------------------------------------------------------------------------
 RegisterNUICallback('dismissPager', function(data, cb)
+    isPagerActive = false
     SetNuiFocus(false, false)
-    local playerPed = PlayerPedId()
-    while not IsEntityPlayingAnim(playerPed, Config.Animations.usePager.dict, Config.Animations.usePager.name, 3) do
-        Wait(100)
-    end
-    StopAnimTask(playerPed, Config.Animations.usePager.dict, Config.Animations.usePager.name, 0.01)
-    if not IsPedInAnyVehicle(playerPed, false) then
-        TriggerServerEvent('96rp-pager:server:PlayAnimation', false)
-    end
+
+    -- play putPagerInPocket and delete pagerObj
+    TriggerServerEvent('96rp-pager:server:PlayAnimation', 'putPagerInPocket')
     Wait(Config.Animations.putPagerInPocket.time)
-    if DoesEntityExist(currentPagerObj) then
+    if DoesEntityExist(currentPagerObj) and not isPagerActive then
         DeleteEntity(currentPagerObj)
     end
     cb('')
@@ -219,17 +232,20 @@ RegisterNUICallback('interactWithContact', function(pagerData, cb)
         TriggerServerEvent('96rp-pager:server:SaveContact', pagerData.value, message.number)
         ShowMessage(message)
     elseif pagerData.interaction == "delete" then
+        -- remove contact from client and server
         local contact = contacts[currentContact]
         table.remove(contacts, currentContact)
 		TriggerServerEvent('96rp-pager:server:RemoveContact', contact.number)
         currentContact = 1
+
+        -- refresh ui and show an other contact if exists
         local contactLeft = nil
         if #contacts > 0 then
             contactLeft = contacts[currentContact]
         end
         ShowContact(contactLeft)
     end
-    cb("test")
+    cb('')
 end)
 
 --------------------------------------------------------------------------
@@ -273,7 +289,7 @@ RegisterNUICallback('showContactRight', function(data, cb)
 end)
 
 --------------------------------------------------------------------------
--- Keyboard interaction for closing pager
+-- Keyboard interaction for closing pager (not working yet)
 --------------------------------------------------------------------------
 RegisterKeyMapping('dismisspager', 'Dismiss a pager', 'keyboard', 'x')
